Part 4/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
SQLAlchemy==2.0.27
python-dotenv==1.0.1
PyJWT==2.8.0
Flask-Mail==0.10.0